style: some clean up
If I understand correctly, the rename_simple_ty thing was used to align
the variable in type with that of the function argument.
For example, it changes the type of \x. e
from
f : (y : Int -> Prop)
to
f : (x : Int -> Prop)

I don't think we rely on the variables in types because we are only
using simple types.
Hence it should be safe to remove it